+++ test-foundry/aave-v3/TestLens.t.sol
@@ -617,7 +617,7 @@ contract TestLens is TestSetup {
         borrower1.supply(aDai, amount / 2);
         borrower1.borrow(aDai, amount / 4);
 
-        (uint256 supply, uint256 borrow, uint256 withdraw, uint256 repay) = morpho
+        (uint64 supply, uint64 borrow, uint64 withdraw, uint64 repay) = morpho
         .defaultMaxGasForMatching();
 
         setDefaultMaxGasForMatchingHelper(0, 0, 0, 0);
@@ -634,7 +634,7 @@ contract TestLens is TestSetup {
         borrower1.supply(aDai, amount / 2);
         borrower1.borrow(aDai, amount / 4);
 
-        (uint256 supply, uint256 borrow, uint256 withdraw, uint256 repay) = morpho
+        (uint64 supply, uint64 borrow, uint64 withdraw, uint64 repay) = morpho
         .defaultMaxGasForMatching();
 
         setDefaultMaxGasForMatchingHelper(0, 0, 0, 0);
