+++ test-foundry/compound/TestLens.t.sol
@@ -880,7 +880,7 @@ contract TestLens is TestSetup {
 
         moveOneBlockForwardBorrowRepay();
 
-        (uint256 supply, uint256 borrow, uint256 withdraw, uint256 repay) = morpho
+        (uint64 supply, uint64 borrow, uint64 withdraw, uint64 repay) = morpho
         .defaultMaxGasForMatching();
 
         _setDefaultMaxGasForMatching(0, 0, 0, 0);
@@ -898,7 +898,7 @@ contract TestLens is TestSetup {
         borrower1.supply(cDai, amount / 2);
         borrower1.borrow(cDai, amount / 4);
 
-        (uint256 supply, uint256 borrow, uint256 withdraw, uint256 repay) = morpho
+        (uint64 supply, uint64 borrow, uint64 withdraw, uint64 repay) = morpho
         .defaultMaxGasForMatching();
 
         _setDefaultMaxGasForMatching(0, 0, 0, 0);
